Added challenge 1 and basic design language.
This is the inital commit of the repository that will contain a list of challenges for new employee's to complete that will test their fundemental understanding of the basics of web dev as well as giving them a chance to learn them at the same time.

For more experienced individuals they should breeze through this and should face no real hurdles, but for those new they should be able to learn on the fly.

The intention is to revise this as issues crop up and there are bits that need to be added.